+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Version 3.1.3
 -------------
 
-Unreleased
+Released 2026-02-18
 
 -   The session is marked as accessed for operations that only access the keys
     but not the values, such as ``in`` and ``len``. :ghsa:`68rp-wp8r-4726`