+++ CHANGES.rst
@@ -27,6 +27,15 @@ Unreleased
     it's disabled in config. Previously, only disabling worked. :issue:`5916`
 
 
+Version 3.1.3
+-------------
+
+Released 2026-02-18
+
+-   The session is marked as accessed for operations that only access the keys
+    but not the values, such as ``in`` and ``len``. :ghsa:`68rp-wp8r-4726`
+
+
 Version 3.1.2
 -------------